Revert "[feat] #44 스프린트2 DetailView 구현"
