+++ ChilledCafe/Data/temparyData.swift
@@ -6,6 +6,14 @@
 //
 import Foundation
 struct constant{
-    let sample = Cafes(name: "꾸꾸하우스", shortIntroduction: "에스프레소 맛에 진심인 카페", thumbnail: "OnboardingImage1", moodImages: ["https://firebasestorage.googleapis.com/v0/b/chilledcafe-1cc36.appspot.com/o/Cafe%2F%EC%98%81%EC%9D%BC%EB%8C%80%2F%EC%95%84%EB%9D%BC%EB%B9%84%EC%B9%B4%EC%BB%A4%ED%94%BC%EB%A1%9C%EC%8A%A4%ED%84%B0%EC%8A%A4%2F1.jpg?alt=media&token=b9a5422a-8b3e-4eec-a350-c4dcc9364313", "https://firebasestorage.googleapis.com/v0/b/chilledcafe-1cc36.appspot.com/o/Cafe%2F%EC%98%81%EC%9D%BC%EB%8C%80%2F%EC%95%84%EB%9D%BC%EB%B9%84%EC%B9%B4%EC%BB%A4%ED%94%BC%EB%A1%9C%EC%8A%A4%ED%84%B0%EC%8A%A4%2F11.jpg?alt=media&token=ad6aaf6d-e13a-4d2c-b290-f7475a7d2b8b", "https://firebasestorage.googleapis.com/v0/b/chilledcafe-1cc36.appspot.com/o/Cafe%2F%EC%98%81%EC%9D%BC%EB%8C%80%2F%EC%95%84%EB%9D%BC%EB%B9%84%EC%B9%B4%EC%BB%A4%ED%94%BC%EB%A1%9C%EC%8A%A4%ED%84%B0%EC%8A%A4%2F1.jpg?alt=media&token=b9a5422a-8b3e-4eec-a350-c4dcc9364313", "https://firebasestorage.googleapis.com/v0/b/chilledcafe-1cc36.appspot.com/o/Cafe%2F%EC%98%81%EC%9D%BC%EB%8C%80%2F%EC%95%84%EB%9D%BC%EB%B9%84%EC%B9%B4%EC%BB%A4%ED%94%BC%EB%A1%9C%EC%8A%A4%ED%84%B0%EC%8A%A4%2F11.jpg?alt=media&token=ad6aaf6d-e13a-4d2c-b290-f7475a7d2b8b",
-                                                                                                                        "https://firebasestorage.googleapis.com/v0/b/chilledcafe-1cc36.appspot.com/o/Cafe%2F%EC%98%81%EC%9D%BC%EB%8C%80%2F%EC%95%84%EB%9D%BC%EB%B9%84%EC%B9%B4%EC%BB%A4%ED%94%BC%EB%A1%9C%EC%8A%A4%ED%84%B0%EC%8A%A4%2F1.jpg?alt=media&token=b9a5422a-8b3e-4eec-a350-c4dcc9364313",], cafeInfo: ["테라스에 앉아 바다를 배경으로 그림 그리기","낮에는 커피를 팔고 밤에는 술을 파는 곳","AR 컨텐츠르로 다양한 공간 스토리를 볼 수 있는 곳"], bookmark: true, ar: true, tag: ["바다와 함께","바다와 함께","AR","AR 경험"], location: "포항시 남구 지곡로 82", businessHour: [""])
+    
+//   static let HotPlaceArray = [
+//        HotPlace(city: "포항", spot: "영일대", imageURL: "spit"),
+//        HotPlace(city: "포항", spot: "효리단길", imageURL: "spit"),
+//        HotPlace(city: "포항", spot: "철길숲", imageURL: "spit"),
+//        HotPlace(city: "포항", spot: "오도리", imageURL: "spit"),
+//        HotPlace(city: "포항", spot: "호미곶", imageURL: "spit"),
+//        HotPlace(city: "포항", spot: "구룡포", imageURL: "spit")
+//        
+//    ]
 }
